requirements.txt
torch
torchvision
torchaudio
torchsde
einops
transformers>=4.28.1
tokenizers>=0.13.3
sentencepiece
safetensors>=0.3.0
aiohttp
accelerate
pyyaml
Pillow
scipy
tqdm
psutil
spandrel
soundfile
kornia>=0.7.1
websocket-client==1.6.3

# Internal API
fastapi==0.112.0
uvicorn==0.30.5
boto3==1.34.158
python-dotenv==1.0.0

# layerdiffuse
diffusers>=0.25.0

# fix for pydantic issues in cog
# https://github.com/replicate/cog/issues/1623
albumentations==1.4.3

# was-node-suite-comfyui
# https://github.com/WASasquatch/was-node-suite-comfyui/blob/main/requirements.txt
cmake
imageio
joblib
matplotlib
pilgram
scikit-learn
rembg

# ComfyUI_essentials
numba

# ComfyUI_FizzNodes
pandas
numexpr

# comfyui-reactor-node
insightface
onnx

# ComfyUI-Impact-Pack
segment-anything
piexif

# ComfyUI-Impact-Subpack
ultralytics!=8.0.177

# comfyui_segment_anything
timm

# comfyui_controlnet_aux
# https://github.com/Fannovel16/comfyui_controlnet_aux/blob/main/requirements.txt
importlib_metadata
opencv-python-headless>=4.0.1.24
filelock
numpy
scikit-image
python-dateutil
mediapipe
svglib
fvcore
yapf
omegaconf
ftfy
addict
yacs
trimesh[easy]

# ComfyUI-KJNodes
librosa
color-matcher

# SUPIR
open-clip-torch>=2.24.0
pytorch-lightning>=2.2.1
